requirements-ml.txt
ultralytics>=8.3
torch>=2.3
torchvision
pillow
numpy
